backend/requirements.txt
Flask==3.0.3
# Data handling & Excel read/write
pandas==2.2.2
openpyxl==3.1.5
# Azure OpenAI client – pinned to a stable version that works with your config.py pattern
openai==1.40.0
httpx>=0.27.0
# Environment variable loading
python-dotenv==1.0.1

# Dashboard charts
matplotlib==3.9.2

# Better date/time handling (helps with Excel serial dates)
python-dateutil==2.9.0.post0
